dev: styles , loaders and seo for subscribers page
