+++ src/lib/components/section/authors/page/user_subscribers/views/FollowersHeader.svelte
@@ -8,11 +8,12 @@
 -->
 
 <script lang="ts">
-  // #region ➤ 📌 VARIABLES
+  import BackButton from "$lib/components/ui/BackButton.svelte";
+  import Tabbar from "$lib/components/ui/Tabbar.svelte";
+  import session from "$lib/store/session.js";
+  import type { IPageAuthorTranslationDataFinal } from "@betarena/scores-lib/types/v8/segment.authors.tags.js";
 
-  import { createEventDispatcher } from "svelte";
-  import DefaultAvatar from "./assets/default-avatar.svelte";
-  import LoggedoutAvatar from "./assets/loggedout-avatar.svelte";
+  // #region ➤ 📌 VARIABLES
 
   // ╭────────────────────────────────────────────────────────────────────────╮
   // │ NOTE:                                                                  │
@@ -26,17 +27,40 @@
   // │ 4. $: [..]                                                             │
   // ╰────────────────────────────────────────────────────────────────────────╯
 
-  export let src: string | null = "",
-    isLoogedIn = false,
-    /**
+  export let author = { name: "Rodrigo Monteirasso" },
+    selection = "subscribers",
+    translations: IPageAuthorTranslationDataFinal;
+
+  const /**
      * @description
-     * avatar size
-     */
-    size = 38;
+     *  📣 `this` component **main** `id` and `data-testid` prefix.
+     */ // eslint-disable-next-line no-unused-vars
+    CNAME: string = "author-profile⮕followers⮕header";
+  $: options = [
+    { id: "subscribers", label: "subscribers" },
+    { id: "followers", label: "followers" },
+    { id: "following", label: "following" },
+  ];
+  $: select = options.find((o) => o.id === selection);
 
-  const dispatch = createEventDispatcher();
-
+  $: ({ globalState, viewportType } = $session);
+  $: isPWA = globalState.has("IsPWA");
+  $: ({ name, username } = author);
   // #endregion ➤ 📌 VARIABLES
+
+  // #region ➤ 🛠️ METHODS
+
+  // ╭────────────────────────────────────────────────────────────────────────╮
+  // │ NOTE:                                                                  │
+  // │ Please add inside 'this' region the 'methods' that are to be           │
+  // │ and are expected to be used by 'this' .svelte file / component.        │
+  // │ IMPORTANT                                                              │
+  // │ Please, structure the imports as follows:                              │
+  // │ 1. function (..)                                                       │
+  // │ 2. async function (..)                                                 │
+  // ╰────────────────────────────────────────────────────────────────────────╯
+
+  // #endregion ➤ 🛠️ METHODS
 </script>
 
 <!--
@@ -49,18 +73,26 @@
 │         │ abbrev.                                                                │
 ╰──────────────────────────────────────────────────────────────────────────────────╯
 -->
-<div class="avatar-wrapper"  on:click={() => dispatch("click")} style="height: {size}px; width: {size}px; {$$restProps.wrapStyle}">
-  {#if src}
-    <div
-      class="avatar-circle"
-      {...$$restProps}
-      style="width: {size}px; height: {size}px;  background-image: url({src}); "
+<div class="wrapper {viewportType}" id={CNAME}>
+  <div class="name-block">
+    {#if !isPWA}
+      <a class="back-button" href="/a/user/{author.usernameLower}">
+        <BackButton mode="back" custom_handler={true} />
+      </a>
+    {/if}
+    <div class="name">{name || username}</div>
+  </div>
+  <div class="tabbar-wrapper">
+    <Tabbar
+      on:select
+      data={options}
+      selected={select}
+      style="gap: {viewportType === 'mobile'
+        ? 40
+        : 24}px; font-size: var(--text-size-m)"
+      {translations}
     />
-  {:else if isLoogedIn}
-    <DefaultAvatar {size} />
-  {:else}
-    <LoggedoutAvatar {size} />
-  {/if}
+  </div>
 </div>
 
 <!--
@@ -74,17 +106,51 @@
 -->
 
 <style lang="scss">
-  .avatar-wrapper {
-    border-radius: 50%;
-    overflow: hidden;
-  }
-  .avatar-circle {
-    width: 38px;
-    height: 38px;
-    border-radius: 100%;
+  .wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+    border-bottom: var(--header-border);
 
-    background-image: url(src);
-    background-repeat: no-repeat;
-    background-size: cover;
+    &.mobile {
+      padding-inline: 16px;
+      gap: 20px;
+      .tabbar-wrapper {
+        margin: auto;
+      }
+
+      .name-block .name {
+        justify-content: center;
+        padding-left: 0;
+      }
+    }
+
+    .name-block {
+      display: flex;
+      justify-content: start;
+      align-items: center;
+      position: relative;
+
+      .back-button {
+        position: absolute;
+        left: 0;
+        top: 0;
+        transform: translateY(-20%);
+      }
+
+      .name {
+        display: flex;
+        color: var(--text-color);
+        justify-self: start;
+        padding-left: 48px;
+        align-items: center;
+        flex-grow: 1;
+        font-family: Roboto;
+        font-size: var(--text-size-l);
+        font-style: normal;
+        font-weight: 500;
+        line-height: 24px; /* 150% */
+      }
+    }
   }
 </style>
